add authenticattion

+++ src/controllers/contact.ts
@@ -4,13 +4,14 @@ import joi from "@hapi/joi";
 import { Contact } from "../model/contacts-mongo";
 import { db, sql } from "../model/contact-postgres";
 
-export async function getContacts() {
+export async function getContacts(decoded:any) {
   // return ContactMongo.find();
-  return db.query(sql`SELECT * FROM  contacts;`);
+  // return db.query(sql`SELECT * FROM  contacts;`);
+  return db.query(sql`SELECT * FROM  contacts WHERE user_id=${decoded.id};`);
 }
 
-export async function getContactByID(contactID: string) {
-  return db.query(sql`SELECT * FROM contacts WHERE id = ${contactID} LIMIT 1`);
+export async function getContactByID(contactID: string, decoded:any) {
+  return db.query(sql`SELECT * FROM contacts WHERE id = ${contactID} AND user_id = ${decoded.id} LIMIT 1`);
 }
 
 type CreateContact = Omit<Contact, "id" | "createdAt" | "updatedAt">;
@@ -36,7 +37,7 @@ const createContactSchema = joi.object<CreateContact>({
   company: joi.string().trim()
 });
 
-export function createContact(contact: CreateContact) {
+export function createContact(contact: CreateContact, decoded:any) {
   const { error, value } = createContactSchema.validate(contact, {
     abortEarly: false,
     stripUnknown: true
@@ -46,22 +47,24 @@ export function createContact(contact: CreateContact) {
     throw error;
   }
 
-  return db.query(sql`INSERT INTO contacts(first_name, last_name, phone, email, company) VALUES(${value.firstName}, ${value.lastName}, ${value.phone}, ${value.email}, ${value.company})
+  return db.query(sql`INSERT INTO contacts(first_name, last_name, phone, email, company, user_id) VALUES(${value.firstName}, ${value.lastName}, ${value.phone}, ${value.email}, ${value.company}, ${decoded.id})
   RETURNING *
   `);
 }
 
-export async function deleteContact(contactID: string) {
-  return db.query(sql`DELETE FROM contacts WHERE id = ${contactID}`);
+export async function deleteContact(contactID: string, decoded:any) {
+  return db.query(sql`DELETE FROM contacts WHERE id = ${contactID} AND user_id = ${decoded.id}`);
 }
 
 export async function updateContact(
   contactID: string,
-  contact: Partial<Contact>
+  contact: Partial<Contact>,
+  decoded:any
 ) {
-  const [existingContact] = await getContactByID(contactID);
+   
+  const [existingContact] = await db.query(sql`SELECT * FROM contacts WHERE id = ${contactID} AND user_id = ${decoded.id} LIMIT 1`)
 
-  if (existingContact === 0) {
+  if (!existingContact) {
     throw new Error("Contact not found");
   }
